Внесла правки
